Add iwl-load, and make sure we disable iwl4965 upon install
We now has two drivers for different kernel releases for
the Intel 4965 device. To account for this we add a script
to allow us to disable and enable one or each other.
By default we disable iwl4965 as we are replacing it with
iwlagn.

If you want to enable iwl4965:

iwl-load iwl4965

If you want to enable iwlagn:

iwl-load iwlagn

By default we'll use iwlagn.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>