compat-2.6: Add place holder for CONFIG_MAC80211_DEBUGFS in config.mk
It is difficult to help someone debug a rate-setting problem when using
compat-wireless because debugging is not set for mac80211. With this patch
the place holder is in the make file even though the debug option is not
set, and it will be easier to coach a user on turning it on.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>