Simply rfkill backport by using macros
By using direct macros for the calls such as:

define rfkill_blocked         backport_rfkill_blocked

we can simpify the backport of rfkill. This works well as
the number of arguments we use will always remain constant
as we are providing our own implemention of rfkill.

This reduces the size of compat.diff, makes it easier to maintain.

Note that compat.diff was also broken as we forgot to use rdev
instead of dev on the core.c patch but since we're removing it
we don't bother to fix it. I note this in case we want to revert
this patch later.

This has been compile-tested on 2.6.27.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>