ar9170 works needs more compat work for 2.6.27
We leave it only enabled for >= 2.6.28,
usb_hcd_unlink_urb() is used within usb_poison_urb()
and although it is available on 2.6.27 its not exported
and cannot be re-implemented. If we figure out a way
to drop the urb from the hardware queue as usb_hcd_unlink_urb()
does then we can backport this.

Also I think we need to backport usb_kill_urb_queue().

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>