Address depmod version differences for depmod.d/demod.conf
http://www.intellinuxwireless.org/bugzilla/show_bug.cgi?id=2061#c6

This adds support for the different demod version out there.

This explanation is taken from the thread which added the patch:

Assume depmod is run without -C or --config option

depmod version 3.6 reads firstly /etc/depmod.conf file and then it
reads config files in directory /etc/depmod.d. Files in directory
/etc/depmod.d are read and parsed in unpredictable order by readdir
function (this function uses raw directory structures of filesystem)

depmod version 3.10 reads /etc/depmod.conf firstly as well, but this
config file is considered as deprecated in this version and prints out
warning about that. Then it reads config files in directory /etc/depmod.d
by readdir function as well, but before it starts to parse them, it sorts
them accord to filename, so you know the order in which they are going to
be parsed and this is very improtant!!! This behavior should be from
version 3.7

The same for both versions:

1) List of directories to be searched is created from search commands from
   all config files in order how particular search commands are found in config
   files (so order of parsed config files is important). If no search command
   is found, updates directory is added before "built-in" directory.

2) As for override command, the last one found for particular module and
   kernel in all config files wins, so again the order of parsed config
   files is important.

From this is evident, that both versions can work differently which depends
on the order of parsed config files from /etc/depmod.d directory that is
different in both versions.

Signed-off-by: Roman Macko <rmacko72@gmail.com>