initial changes for handling compat.diff
We change the way we deal with backport patches.
We used to just apply compat/compat.diff but this
can grow in size as we move on through kernels and
maintaining can be painful. To ease with its
maintenance and review of patches we add add
the patches onto their own directory and separate
them based on the functional nature of the patch.

We add 3 files for now:

01-netdev-ops.patch
02-ksize.patch
03-rfkill.patch

This reflects better what each patch addresses.

While at it update the hunks necessary to apply
cleanly as of wirless-testing master-2009-08-04.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>