Move netns hunk to its own 05-netns.patch
There is another netns related change on 01-netdev.patch
but haven't figured out a way to split by having them
not touch the same files. While we could have the netdev
stuff apply first and then afterwards apply the netns
stuff to the same files maintaining that would be quite
painful so for now lets leave this as is.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>