We ensure disable mac80211 mesh by relabelling it
Mesh is currently busted.

If your kernel has an option enabled we cannot
simply disable it for two reasons:

* kconfig magic
* autoconf

We cannot disable CONFIG_MAC80211_MESH by
using undef on linux/compat_autoconf.h,
and even if it goes after linux/autoconf.h.

Even if you solve that you still need a way
to unset it from the kconfig somehow.

Mesh is currently busted and disabled so even
if your kernel *does* enable it we want to disable
it for you. We do this by relabelling the mesh
kconfig option.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>