compat-2.6: uncomment out the wiphy_type on older kernels
Kernels >= 2.6.32 can identify the type of device netdevice
so that sysfs can be used to get this. We never really had a
systematic way of doing this -- now we do through the
SET_NETDEV_DEVTYPE() macro. For older kernels we make the
SET_NETDEV_DEVTYPE() be a no-op this means the wireless type
we define is unused so we ucomment it simply to avoid a
compile warning.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>