Fix rfkill backport on older kenrels
Although the rfkill miscdevice on older kernels would end up
being associated with a diferent class there is some wierd
issue which causes an oops on 2.6.30 if rkfill is enabled
and two devices are created in competition for the first
rkill%d device name.

This happens even with old rfkill on the "rfkill" class
and backport_rfkill on the "backport_rfkill" class.

This was tested to be the fix required for 2.6.30.

Tested-by: Greg KH <gregkh@suse.de>
Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>