Use linux-next instead of two separte trees
No need to use two separate trees, so just use linux-next
for both 802.11 updates and bluetooth updates.

What helps also is we'll be now using wireless-next for a clean
wireless history for pull requests for some drivesr that choose
to send pull requests and updates do go to linux-next for 802.11
updates in at about 1-2 days.

Bluetooth pushes changes to linux-next on a less regular basis due
to the smaller amount of changes kept track of on that tree.

What this means to users of compat-wireless is they'll get more
ready code being tested and used. The simplicity of maintaining
a tree with only one tree for upstream code outweighs the benefit
of getting the few random extra patches not yet queued up in
linux-next.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>