Let this thing build without requiring CONFIG_BT=m
The Bluetooth port is separate from the Wireless port
so don't require CONFIG_BT=m to build this. We can do the
same for wireless but for now this at least lets Ubuntu
users build this thing -- Ubuntu currently ships with
CONFIG_BT=y from 2.6.27..2.6.31.

If we want to let compat-blutooth be available to Ubuntu
users on Ubuntu Lucid this config option will need to be
revised. The CONFIG_BT just builds net/bluetooth/bluetooth.c,
that's all -- so it seems worth it. Fedora is shipping with
CONFIG_BT=m , not sure about other distributions.

This also fixes some final touches to get the build
going again which I missed in my previous patches.

./scripts/drivers-select is likely very busted now.
We'll need to fix this.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>