compat-wireless: modules declared as PHONY target in Makefile
When a particular file alone is modified in compat package, it is not compiled with make.
To compile the modified file compat package needs to be cleaned up and all files need to be compiled.
This is because all files are compiled in modules target, and file by name modules is not generated.
So this target will be considered up-to-date once compiled.

modules target is declared as PHONY target to avoid this problem.

I feel config.mk need not be included when KERNELRELEASE is not NULL as
it degrades the performance and doesn't make any difference.  Please comment.

Signed-off-by: Bala Shanmugam <sbalashanmugam@atheros.com>