Backport pm-qos for kernels <= 2.6.24
pm-qos is used by mac80211 and ipw2100. mac80211 uses it for
the dynamic power save to see if to enable it or not upon
recalculation for dynamic ps. It takes the latency registered
to the kernel into account.

Since older kernels do not have pm-qos we simply initialize
pm-qos stuff durng cfg80211 init as if it were part of cfg80211.
On future kernels we can just do this in compat.ko

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>