Add support for driver-select to enable older kernels
compat-wireless currently works as a whole down to 2.6.25.
The excemption to this is ar9170 which cannot yet be backported
due to the urb anchor poisoning which we haven't figured out
how to back port (not sure if its even possible). Sometimes
though we want to push backports down even further. In those
cases we don't want to have to force ourselves to port
the entire package down but maybe just a single driver.

For that we add support for through the files

enable-old-kernels/enable-2.6.2x.patch

Force the patch to be applied on the driver select
for your driver. We now enable ath5k and ath9k on
2.6.24.

Careful -- if you run driver-select on ath5k or ath9k
make sure to do:

git checkout config.mk
git checkout Makefile
git checkout scripts/gen-compat-autoconf.sh

before commiting something on your git tree if
you are developing. Unless of course you know what
you are doing.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>