compat-wireless: backport netdev_queue usage for kernels <= 2.6.27
The patch:

commit 79d16385c7f287a33ea771c4dbe60ae43f791b49
Author: David S. Miller <davem@davemloft.net>
Date:   Tue Jul 8 23:14:46 2008 -0700

   netdev: Move atomic queue state bits into netdev_queue.

   Signed-off-by: David S. Miller <davem@davemloft.net>

Allowed allowed for the state of each queue to be separated.
Previoust to this all queue information relied on an internal
central state to the netdev. To backport calls to start/stop
each netdev queue then we just need to stop/start all queues
unfortunately. This should have a performance impact when
the device is gets a lot of congestion.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>