Fix MQ support for 2.6.23..2.6.26
Turns out the 2.6.23..2.6.26 kernels did have MQ support
but relied on the notion of a struct net_device_subqueue
instead of a struct netdev_queue. For these kernels we just
do a direct mapping of calls, for kernels older than 2.6.23
we start/stop all queues whenever a specific queue is mapped.

More review is required on the select_queue stuff. I'm not
sure yet if how we implement the backport we are ensuring
we queue the skb onto the right queue.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>