Add linux-next cherry picked patches directory
We work hard to get patches in time onto the stable
tree but sometimes a few things slip out, and sometimes a
stable fix is simply too big in size to be merged into
stable. In such cases though we do believe some of these
patches are still relatively important to either enable new
hardware which escaped the late rc cycles or to correct some
behaviour which might be too late for stable. We apply
these patches by default as they will be supported on these
releases.

The larger the number of patches you see in this new directory
the more we should be ashamed. We should strive to reduce this
to 0 all the time.

We start by adding Atheros AR8152 and AR8152 support through
atl1c, the patch is already merged into linux-next:

	Subject: [PATCH] atl1c: Add support for Atheros AR8152 and AR8152

	AR8151 is a Gigabit Ethernet device. AR8152 devices are
	Fast Ethernet devices, there are two revisions, a 1.0
	and a 2.0 revision.

	This has been tested against these devices:

	Driver  Model-name      vendor:device   Type
	atl1c   AR8131          1969:1063       Gigabit Ethernet
	atl1c   AR8132          1969:1062       Fast Ethernet
	atl1c   AR8151(v1.0)    1969:1073       Gigabit Ethernet
	atl1c   AR8152(v1.1)    1969:2060       Fast Ethernet

	This device has no hardware available yet so it goes untested,
	but it should work:

	atl1c   AR8152(v2.0)    1969:2062       Fast Ethernet

	Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>