Add linux-next cherry picked patches directory
We work hard to get patches in time onto the stable
tree but sometimes a few things slip out, and sometimes a
stable fix is simply too big in size to be merged into
stable. In such cases though we do believe some of these
patches are still relatively important to either enable new
hardware which escaped the late rc cycles or to correct some
behaviour which might be too late for stable. We apply
these patches by default as they will be supported on these
releases.

The larger the number of patches you see in this new directory
the more we should be ashamed. We should strive to reduce this
to 0 all the time.

This only makes sense for the stable compat-wireless releases
as bleeding edge compat-wireless already uses linux-next. We
add this patch into this tree though so that future stable
branches will get it as well.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>