mac80211: use netdev_attach_ops()
This simplifies mac80211's net_device_ops's backports
significantly. An unexpected benefit was the simplification
of the MQ backport, this is possible since the select_queue
callback can now simply be bacported by netdev_attach_ops()
selectively depending on the kernel used, the callback itself
can always be set on the ops structure since even older kernels
will have it on the code, it will just be unused.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>