Use the defines to tag the compat module
The compat module now has read-only module parameters which
we can help tune to help identify the exact upstream base
tree used for this backport, as well as the specific tree
used to make it. For example when building a stable
compat-wireless release based on the linux-2.6-allstable.git
git tree, using the tree as of the 'v2.6.34' tag we'd get:

$ sudo grep '' /sys/module/compat/parameters/compat_
/sys/module/compat/parameters/compat_base_tree:linux-2.6-allstable.git
/sys/module/compat/parameters/compat_base_tree_version:v2.6.34
/sys/module/compat/parameters/compat_version:v2.6.34-rc4-4-g9ace871

The last file indicates compat-wireless.git was used, but
specifically that the v2.6.34-rc4 branch was used and that
4 patches have been applied on top of that tree. The top
sha1sum commit on that tree is 9ace871.

These files will work even if you have built compat-wireless
completely as built-in to your kernel.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>