compat-wireless: add the Kconfig for drivers
The linux-next-chery-pick, linux-next-pending, and crap
patches all would rely on the Kconfig so add the Kconfig
files internally.

We could eventually use these as well to modify the
build system for compat-wireless so we can use mconf
intsead. For that we'd need to modify each kconfig option
though with some heroic effort.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>