compat-wireless: fix backport code calculation
The current computation was not taking into account the
compat/*.[ch] files and the include/linux/compat*.h files.
Additionally these files were just being counted towards the
total sum of the upstream code.

Remove the compat code from the total sum of upstream code,
and add the compat code to the backport sum. This also adds
a new line to indicate to us the size of the total backport
work (patches + compat) and compat alone.

These are the stats I get today:

compat-wireless code metrics

    491620 - Total upstream code being pulled
      1393 - backport code changes
      1163 - backport code additions
       230 - backport code deletions
      5679 - backport from compat module
      7072 - total backport code
      1.44 - % of code consists of backport work

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>