compat-wireless: use /etc/init.d/ and not sudo service
sudo is not installed and used on all distributions. the user should
call the script with the needed rights and the script should not
include sudo. This fails on debian lenny.

service is only available on very recent distributions and not on old
ons like debian lenny in the default configuration. Use the old init.d
script. This also works on ubuntu 10.04.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>