compat-wireless: update 28-pm-qos-params.patch
This patch is required because now pm-qos has 3 different flavors:

On kernels >= 2.6.36 the pm qos req is kzalloc()'d by pm-qos itself
On kernel   = 2.6.35 the pm qos req uses the stack
On kernels <  2.6.35 the pm qos req is pegged to a unique name

On newer kernels we don't use a name to tag a pm-qos requirement, and on
older kernels we do. Although we can port over the struct pm_qos_request_list
to older kernels it provides no variable with any unique name we can use
for older kernels. Perhaps there's a way.. right now I just don't see it.

We also should probably not bother updating pm-qos to newer versions unless
we are willing to treat the oldest kernels as newer ones and that'd be a
little wierd. Update the header of 28-pm-qos-params.patch while at it
with a little explanation as to why this is all needed.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>