compat-wireless: add ifdefs into compat_autoconf.h
When using ifdef or ifndef in config.mk these conditions are now also
written into compat_autoconf.h. This replaces some special handling in
gen-compat-autoconf.sh. This was needed because CONFIG_RT2X00_LIB_LEDS
should only be selected for kernel > 2.6.24.
The script adds ifdefs, ifndefs else and endif into the header file. It
adds endif and else only if it is followed by " #CONFIG_" This is
needed because we do not support something like "ifneq
($(CONFIG_CRC7),)" or "ifeq ($(wildcard $(KLIB_BUILD)/.config),)" for
now.
Now egrep is needed because grep does not support the needed regular
expressions.
The spaces in the egrep part have to be replaced by some other
character, a + is used here.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>