compat-wireless: rename CONFIG_IWLWIFI config symbol
In kernel 2.6.24 CONFIG_IWLWIFI was a bool config symbol and not a
tristate one. In that kernel it was used for showing the menu option for
the Intel driver and not for selection the module it is used for today.
In kernel configurations for version 2.6.24 it is set so y and not to m.
As we do not overwrite some value selected in the kernel config it
will cause problems when we build some parts of the Intel driver in
compat-wireless directly into the kernel.
Now we use an other name for this config symbol to work around the
problem.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>