compat-wireless: remove special handling for CONFIG_CFG80211_WEXT
The special handling for CONFIG_CFG80211_WEXT is not needed any more,
because it is only activated when CONFIG_WIRELESS_EXT is also set.
Enabling this condition only for kernel <= 2.6.32 was wrong. In compat-
wireless we are able to deactivate CONFIG_WIRELESS_EXT for every
kernel. For activating it we need CONFIG_WIRELESS_EXT on every kernel.
If kernel <= 2.6.31 was build with cfg80211 support CONFIG_WIRELESS_EXT
was also selected, so it was activated in most cases for that kernel
versions. But it is possible to build compat-wireless without having
cfg80211 while building the kernel and that caused problems in the past.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>