compat-wireless: fix unresolved symbols with old kernels
CONFIG_WEXT_PRIV and CONFIG_WEXT_SPY were introduced with kernel 2.6.33
and are not available in older kernel versions. We should just depend on
them for recent kernel versions, but this feature is not implemented in
compat-wireless yet. If compat-wireless was commpield against kernel <
2.6.29 with e.g. CONFIG_IPW2200=m then libipw was not build in compat-
wireless but ipw2200 was build, because the config options set in the
kernel config were used. For now we build ipw2*00 every time
CONFIG_WIRELESS_EXT is enabled. This could cause some problems with
kernel > 2.6.32 and CONFIG_WEXT_PRIV=n but there are no such
distribution kernel available and we haven't got an bug report because
of this problem although is problem existed for a long time.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>