compat-wireless: add CONFIG_ATH9K_RATE_CONTROL=y
Disable this to get minstrel as default, we leave the ath9k
rate control algorithm as the default for now as that is also
default upstream on the kernel. We will move this to minstrel
as default once we get minstrel properly tested and blessed by
our systems engineering team. CCK rates also need to be used
for long range considerations.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>