compat-wireless: get pending stable patches
We are close to the merge window but there 23 pending
stable fixes that haven't made it to Linus' tree.
Instead of waiting lets suck them them into compat-wireless
into a new directory so users can take advantage of them.

This is a direct copy of what -s generates, nothing more,
next we'll refresh the patches. For future updates we simply
do:

./scripts/admin-update.sh -s refresh

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>