compat-wireless: update code-metrics to reflect extra patches applied
This will peg the characters options we used to apply extra
patches to our compat-wireless release to identify the extra
patches applied.

For example with "-s", we will now get:

compat-wireless-v2.6.36-rc5-1-2-g2ac99dc-s

This should be also picked up on the compat module which when loaded
will display the above string. This is useful when troubleshooting
issues and trying to find out what a user is using exactly.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>