compat-wireless: Added ath3k sflash workaround to crap directory
This patch in crap directory enables btusb to load firmware
to device RAM when it is plugged in.

Reason for this not getting upstream:

It was first rejected without any alternatives suggested. Then
an alternative was finally suggested but too late for integration
upstream.

This is a work around until a new firmware and patch is being worked
on which is acceptable upstream.

For details refer to:

http://lkml.org/lkml/2010/10/5/195

Signed-off-by: Bala Shanmugam <sbalashanmugam@atheros.com>