compat-wireless: fix activesync devices
Compiling support for ActiveSync devices in compat-wireless depends
on USB_NET_CDCETHER to be set in main kernel.

But USB_NET_CDCETHER can also be a module, in which case
CONFIG_USB_NET_CDCETHER will not be defined inside compat_autoconf.h and
the compat-wireless cdc_ether.ko will miss the support for it.

So check also for CONFIG_USB_NET_CDCETHER_MODULE in config.mk to be safe
in both cases.

Signed-off-by: Sven-Haegar Koch <haegar@sdinet.de>