compat-wireless: backport stable patch: orinoco-initialise-priv-hw-before-assigning-the-inte.patch
The stable patch:

orinoco-initialise-priv-hw-before-assigning-the-inte.patch

requires a manual backport because the hunks in the upstream patch
affect areas in compat-wireless which were already touched by code
in the patches/ directory, so the upstream hunk breaks. So the
code compat-wireless for orinoco already had some ifdefs over
the kernel version.

We are going to make the wild assumption this patch works for
older kernels too... if it busts please report!

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>