compat-wireless: fix stable pending tag mismatch output
If Linus made an update to his tree, say a 2.6.38-rc3 release,
but we have not yet updated our linux-next.git tree the tag will
not be there and we will not see the failure. I add some check now
to ensure the tag is present, otherwise we bail out and tell you
exactly the command we used and the directory it was run under.

Signed-off-by: Luis R. Rodriguez <lrodriguez@atheros.com>