compat-wireless: support RHEL6 as a build target
This patch enables compatibility with RHEL6 as a build target.  This
mostly includes changes to the configuration system.

Infrastructure is added to config.mk to provide a series of RHEL
compatibility definitions (e.g. CONFIG_COMPAT_RHEL_6_1).  A couple of
checks are added to set appropriate config options for RHEL6 kernels.
Finally, gen-compat-autoconf.sh is modified to correctly handle the
CONFIG_COMPAT_RHEL_* usage in config.mk.

Signed-off-by: John W. Linville <linville@tuxdriver.com>
Signed-off-by: Luis R. Rodriguez <mcgrof@qca.qualcomm.com>