compat-wireless: add CONFIG_COMPAT_BT_SOCK_CREATE_NEEDS_KERN
An API change in the 2.6.33 timeframe made a number of bluetooth
functions require an extra parameter.  Some kernels (e.g. RHEL6)
backport that API change, and this allow us to accommodate them.

This adds CONFIG_COMPAT_BT_SOCK_CREATE_NEEDS_KERN to config.mk and
modifies an existing patch for the Bluetooth subsystem.

Signed-off-by: John W. Linville <linville@tuxdriver.com>
Signed-off-by: Luis R. Rodriguez <mcgrof@qca.qualcomm.com>