compat-wireless: build: Throw error if space in path
Because of make limitations, the build will fail if there are spaces in
the path leading to the compat-wireless directory.
I tried to make it work anyway for a while and, while it probably is
possible, it would involve rewriting part of kbuild and well...
This at least throws up nicely if there is a space in the path.

Signed-off-by: Dominique Martinet <asmadeus@codewreck.org>
Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>