compat-wireless: check if modprobe exists before using it
If you're cross-compiling compat-wireless on a different host, you
can't reliably use modprobe.  The host system may not even have
modprobe, so it doesn't make sense to run modprobe, unless you're
installing the modules on the target machine.

The modprobe command was being used unconditionally to set the MADWIFI
and OLD_IWL variables.  This commit adds a check to see whether
modprobe exists before trying to use.  It also adds a dependency on
the modprobe commands to all recipes that rely on it, so we fail more
gracefully if modprobe doesn't exist in the system.

Reported by: William List <wlist@ti.com>
Signed-off-by: Luciano Coelho <coelho@ti.com>