compat-wireless: fix patches/99-change-makefiles.patch
CONFIG_USB_NET_QMI_WWAN was added and therefore the
respective Makefile, drivers/net/usb/Makefile, needs
to be adjusted to remove that from our list of targets.

If we user our own generated Makefiles for drivers the
advantage is that we would not face this issue when upstream
gets a new driver that our hunk needs to ignore but -- the
disadvantage is that any changes upstream would require a
respective change into compat-wireless. Driver Makefiles
though typically are just copied as-is and never modified.
Its subsystem Makefiles that we tend to have to patch on
this file.

One option would be to somehow annotate upstream targets as
compat-wireless maintained so that we can script out what
targets to leave in.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>