compat-wireless: remove recurssive call on building modules
The recursion call to build modules is not required. It was added
due to the fact that we now included a new file, CONFIG_COMPAT
and we needed it at build time. We do not need the recursive call
given that the kernel will already call the same Makefile later
when building modules. We take advantage of this fact and simply
ensure we build CONFIG_COMPAT when processing our rules.

We do not define a direct target for CONFIG_COMPAT given that
we want to force a regeneration of the file and compat autoconf,
whenever the build environment changes. We still add a dummy
CONFIG_COMPAT target rule to avoid Make looking rules for it.

In order for this build to work we now just need to export
CFLAGS so that the variables defined on the first pass are
available later at module build time.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>