compat-wireless: fix collateral damange of cherry picking
We cherry picked in the -n release the patch,

linux-next-cherry-picks/0007-mac80211-handle-non-bufferable-MMPDUs-correctly.patch

which apart from some things also renamed IEEE80211_TX_CTL_POLL_RESPONSE.
The linux-next change did not address carl9170 likely as it did not
have that line present there but in older kernel code we do have it.
Fix this by ammending the patch and adding the respective change
for carl9170.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>