compat-wireless: backport assinging the tty dev to hci dev
This backports assinging the tty device to the hci parent
pointer. The goal was to make the devices link under /sys/
but on older kernels we cannot support this. One alternative
that occurs to me is to have a compat subsystem where userspace
applications can hunt for such newer devices, but this requires
quite a bit of work and consideration.

commit 7f4b2b04c88377af30c022f36c060190182850fb
Author: Andrei Warkentin <andreiw@motorola.com>
Date:   Fri Feb 11 17:19:26 2011 -0600

    Bluetooth: Make hci a child of the corresponding tty device.

    Make /sys/class/bluetooth/hciX a symlink to
    path under corresponding tty.

    Signed-off-by: Andrei Warkentin <andreiw@motorola.com>
    Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>