compat-wireless: license under ISC files for standalone builds
This licenses under the ISC license scripts/refresh-compat,
scripts/refresh-compat-local and the top level Makefile to help
standalone development code be ISC licensed. The goal is to help
a standalone development tree to be used to target either Linux
or BSD.

In the Linux target case although the files may be ISC licensed, as it
was discussed on lkml [0] the run time license of the module will
still be GPL, so no one should even think that they can make proprietary
or BSD-licensed only modules out of the Linux kernel with this framework.

The goal is to let a standalone git tree exist and allow modules to be
build either for Linux or BSd.

[0] http://article.gmane.org/gmane.linux.kernel/1278784

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>