compat-wireless: add support for CONFIG_HID_GENERIC
The HID driver code was moved to a separate module under
CONFIG_HID_GENERIC. We remove the old code that had patched
support for it given that ifdef'd code around calls to
hid_register_driver() and friends given that this was only
available on kernels >= 2.6.28. We simply only compile the
code in question on newer kernels.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>