compat-wireless: remove master-tag file
The master-tag file was used to describe from where we pulled
code from when wireless-testing was used instead of linux-next.git.
It turns out the way this was being done could lead to incorrect
descriptions given that the master-tag file was generated by
looking at all the git tags present and using the latest master
tag. The assumption is that your tree would not have any newer
master tags. Turns out that we can use git describe to extract
correctly the proper master tag, at least now this seems to work
well on wireless-testing. Even if it does not work the git describe
tag is a better representation and would lead to less mistakes.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>