compat-wireless: address usage of request_module_nowait() on iwlwifi
request_module_nowait() was added on 2.6.30 but we can't
backport it. Under some circumstances though request_module_nowait()
likely was not meant to be used on purpose and instead we can
assume request_module() will work. This patch addresses the few
places where request_module_nowait() was used but that we can
at least somewhat assume that this can work. We do not backport
this into compat.git as a direct static inline given that there
are surely some places where request_module() cannot be used.

All users of request_module_nowait() should revise their code and
snsure that is indeed what was meant.

In this patch we address using the regular request_module() on
iwlwifi given that it seems we can use it there. Someone can
hopefully however test and verify this is OK. For newer kernels
that have request_module_nowait() we leave the call.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>