compat-wireless: fix stable alias on admin-update.sh
When running ./scripts/admin-update.sh -s -- the code will look
for all pathces on linux-next since the current head tag on the
stable release we are on that have the stable@kernel.org
on the commit log. The e-mail address alias has changed
to stable@vger.kernel.org. This patch updates that alias so
we can continue to run:

./scripts/admin-update.sh -s

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>