compat-wireless: move alx pending patches to crap
The alx driver needs to remove the old devices that atl1c
already supports. This work is being done but users can
still use the alx driver. To help these users move the
alx patch to crap/ and add a new patch that disables all
the atl1c devices from the alx driver. This means we no
longer need a script to disable atlc.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>