compat-wireless: add ath9k linux-next-pending patches
These patches are posted on the linux-wireless mailing list
but are not yet merged into linux-next.git. The QCA team
has expressed interest in getting these merged. This consists
of the usage of the new ANI on pre-AR9003 chipsets as well
as some few driver fixes recently posted which are even
marked as stable.

The patch:

linux-next-pending/0004-ath9k-fix-invalid-pointer-access-in-the-tx-path.patch

required a bit of massaging to port it to v3.5-rc3.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>